fix: remove contact info app and move to profiles
