--- contact_information/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-import os
-
-
-class ContactInformationConfig(AppConfig):
-    name = os.path.basename(os.path.dirname(os.path.abspath(__file__)))
