+++ profiles/forms.py
@@ -1,6 +1,6 @@
 from django import forms
 from django_ckeditor_5.widgets import CKEditor5Widget
-from .models import Summary
+from .models import Summary, ContactInformation
 
 
 class SummaryForm(forms.ModelForm):
@@ -17,7 +17,26 @@ class SummaryForm(forms.ModelForm):
         }
 
         widgets = {
-              "summary": CKEditor5Widget(
-                  attrs={"class": "django_ckeditor_5"}, config_name="default"
-              )
-          }
+            "summary": CKEditor5Widget(
+                attrs={"class": "django_ckeditor_5"}, config_name="default"
+            )
+        }
+
+
+class ContactInformationForm(forms.ModelForm):
+    """
+    Form to create contact information.
+    """
+    class Meta:
+        model = ContactInformation
+        fields = [
+            'first_name', 'last_name', 'email', 'phone_number',
+            'address1', 'address2', 'city', 'zip_code', 'country',
+            'linkedin', 'github'
+        ]
+        labels = {
+            "address1": "Address Line 1",
+            "address2": "Address Line 2",
+            "github": "GitHub URL",
+            "linkedin": "LinkedIn URL"
+        }
