+++ profiles/templates/profiles/profile.html
@@ -72,7 +72,7 @@
 
 {% endblock %}
 
-{% block extra_js  %}
+{% block extra_js %}
 {{ block.super }}
 
 <script>
@@ -121,8 +121,8 @@
 
     // instance options with default values
     const instanceOptions = {
-    id: 'tabs-example',
-    override: true
+        id: 'tabs-example',
+        override: true
     };
 </script>
 {% endblock %}
